Few fixes: use json-c for json parsing (portable), fix type problem with lcmapsd_timeout, update README to have correct timeout value.

+++ nss_lcmapsd.h
@@ -16,7 +16,7 @@
 typedef struct lcmapsd_opts_s	{
     char *lcmapsd_conffile;
     char *lcmapsd_url;
-    int lcmapsd_timeout;
+    long lcmapsd_timeout;
 } lcmapsd_opts_t;
